requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
flask-login==0.6.2